requirements.txt
# Core dependencies
arxiv==1.4.7
PyPDF2==3.0.1
openai==1.3.0
pyyaml==6.0.1
requests==2.31.0
